Refactor models => Payload
