+++ apps/charterafrica/src/lib/payload/constants.js
@@ -1,0 +1,4 @@
+export const ORGANIZATION_COLLECTION = "organisations";
+export const CONTRIBUTORS_COLLECTION = "contributors";
+export const TOOL_COLLECTION = "tools";
+export const ECOSYSTEM_GLOBAL = "ecosystem-schema";
