+++ apps/charterafrica/src/payload/collections/Contributors.js
@@ -1,4 +1,4 @@
-import { CONTRIBUTORS_COLLECTION } from "../../lib/ecosystem/models";
+import { CONTRIBUTORS_COLLECTION } from "../../lib/payload/constants";
 import avatarUrl from "../fields/avatarUrl";
 import dateField from "../fields/dateField";
 import slug from "../fields/slug";
