+++ apps/charterafrica/src/payload/collections/Tools.js
@@ -5,7 +5,7 @@ import {
 import {
   TOOL_COLLECTION,
   CONTRIBUTORS_COLLECTION,
-} from "../../lib/ecosystem/models";
+} from "../../lib/payload/constants";
 import avatarUrl from "../fields/avatarUrl";
 import dateField from "../fields/dateField";
 import slug from "../fields/slug";
