+++ apps/charterafrica/src/payload/globals/Ecosystem.js
@@ -1,4 +1,4 @@
-import { ECOSYSTEM_GLOBAL } from "../../lib/ecosystem/models";
+import { ECOSYSTEM_GLOBAL } from "../../lib/payload/constants";
 import airtableBaseSelect, {
   schema,
 } from "../fields/ecosystem/airtableBaseSelect";
